updating folder structure
