+++ js/constants/constants.js
@@ -1,0 +1,10 @@
+export const stationNames = {
+  "Euston Station": "940GZZLUEUS",
+  "Finsbury Park Station": "940GZZLUFPK",
+  "Vauxhall Station": "940GZZLUVXL",
+  "Green Park Station": "940GZZLUGPK"
+};
+
+export const inboundTable = document.querySelector("#inbound-table");
+export const outboundTable = document.querySelector("#outbound-table");
+export const form = document.querySelector("form");
