+++ js/utils/handleNoInfo.js
@@ -1,0 +1,24 @@
+import { inboundTable, outboundTable } from "../constants/constants.js";
+import { clearTable } from "./clearTable.js";
+
+export const handleNoInfo = (direction) => {
+  console.log("No departures for " + direction + " at this time");
+
+  const currentTable = direction === "inbound" ? inboundTable : outboundTable;
+
+  // remove thead and tbody
+  const tableHead = currentTable.querySelector("thead");
+  const tableBody = currentTable.querySelector("tbody");
+  tableHead.remove();
+  clearTable(tableBody);
+
+  // add error message to the current table
+  const row = document.createElement("tr");
+  // set the class of row to be error_message
+  row.classList.add("error-message");
+  const cell = document.createElement("td");
+  cell.textContent = "No departure info at this time";
+  row.appendChild(cell);
+  tableBody.appendChild(row);
+  return;
+}
