+++ tfl.html
@@ -107,7 +107,7 @@
         </tbody>
       </table>
     </section>
-    <script src="js/tfl.js"></script>
+    <script src="js/tfl.js" type="module"></script>
     <script src="https://kit.fontawesome.com/7a65a08da4.js" crossorigin="anonymous"></script>
   </body>
 </html>
